DeleteNode Seg Fault Error Fix
I found, and fixed the error in the linked list code that was causing the seg faults. I also left a comment in the source code to denote the logical error that was causing the faults. Everything compiles correctly, except 1 thing:

When I startt the project and enter the grid size I get a notification:

"Error directory not found"

The only other output is as follows:

QObject::connect: No such slot QSpinBox::on_sbox_ValueRed_editingFinished() in ..\GoofyLightsEditor\mainwindow.cpp:67
FrameList testing
Program end

So the linked list should now be good to go.